+++ src/lib/config.ts
@@ -1,0 +1,6 @@
+
+const Config = {
+  SCAN_COUNT: 100,
+};
+
+export default Config;